Snap: Enlarge the saving queue size to 60M
The picture would not be saved if the saving queue is full. Currently,
the queue is very easy to reach 50M during continuous shot with the
13M pixel pictures, that caused many pictures lost during continuous
shot.

Change the saving queue max size to 60M will avoid this issue,
mean while change the LOW_STORAGE_THRESHOLD_BYTES as the same value.

CRs-Fixed: 851554

Change-Id: Ib93f05404371872f2b82d489378c01a6122ccb6b
Signed-off-by: Zdrowy Gosciu <ZdrowyGosciu+GITHUB@gmail.com>