Snap: Set preview size to 1280 x 720 when 720p is used
Some high-speed modes supported at 720p are not supported at higher resolutions,
so the camera preview would become choppy when selecting high-speed modes supported
by the video resolution with a camera preview resolution set that does not support
them.

e.g. 120 FPS is supported at 720p, but not 1080p, so the preview size in this
case must be 720p.